-There was some logic blocks missing for the two endpoint in the goal system. Whilst testing I found out that it doesn't verify if the goals that it wants to get or delete are in the repository

+++ MoneyOrbit/MoneyOrbit/Infrastructure/Services/GoalService.cs
@@ -110,7 +110,10 @@ namespace MoneyOrbit.Infrastructure.Services
         public async Task<ResultObject> GetGoal(GetGoalDto getGoalDto)
         {
             var goal = await GetGoalfromID(getGoalDto.GoalID);
-            //Returns the amount accomplished by the goal
+            //Check if the goal exists in the database
+            var goalquestion = await _goalRepository.GetInstanceOfType<Goal>(getGoalDto.GoalID);
+            if (NullGuard.IsNull(goalquestion))
+                return new ResultObject() { Error = "The provided ID does not exist in the goal repository." };
             return new ResultObject() { Result = goal };
         }
         public async Task<ResultObject> GetGoalAmountAccomplished(GoalAmountAccomplishedDto gAADto)
@@ -180,7 +183,10 @@ namespace MoneyOrbit.Infrastructure.Services
                 return new ResultObject() { Error = "Goal delete data is null." };
             if (String.IsNullOrEmpty(dGDto.GoalID))
                 return new ResultObject() { Error = "Goal ID is required to update a goal." };
-
+            //Check if the goal exists in the database
+            var goalquestion= await _goalRepository.GetInstanceOfType<Goal>(dGDto.GoalID);
+            if(NullGuard.IsNull(goalquestion))
+                return new ResultObject() { Error = "The provided ID does not exist in the goal repository." };
             await _goalRepository.DeleteData(dGDto.GoalID);
 
             return new ResultObject() { Result = "success" };
